add first type annotation to driver.interface
I'm planning on annotating as much as is needed for intellisense tools to grasp just about what's going on.
Because Pylance can't infer the type of objects due to the scarcity of type annotations,
which results in a lot of functions not having hover summaries, meaning I have no clue what they do.
Having these would make it easier for anyone who's reading the testbench file after, not have to also dig through the source files like I am.